Pictures of shock shock interaction
To implement this, the equations were non dimentionalised, and the initial condition was changed. This allowed for larger time steps. In these pics, the flow velocity is at 1000m/s while free stream speed of sound is 347m/s, which converts roughly to inflow mach of 2.88. The density shown here is non-dimentionalised density.